- Finish "ex7"

+++ ex7/VietHQ_ASM_FRAMEWORKS_NGAY_7/src/main/java/com/viethq/shoppingonline/controllers/CategoryController.java
@@ -1,22 +1,34 @@
 package com.viethq.shoppingonline.controllers;
 
 import com.viethq.shoppingonline.entities.Category;
+import com.viethq.shoppingonline.entities.Product;
 import com.viethq.shoppingonline.repositories.CategoryRepository;
+import com.viethq.shoppingonline.repositories.ProductRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.validation.BindingResult;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.servlet.ModelAndView;
 
+import javax.validation.Valid;
+import java.util.List;
+
 @Controller
 @RequestMapping("/category")
 public class CategoryController {
 
     private CategoryRepository categoryRepository;
+    private ProductRepository productRepository;
+
 
     @Autowired
     public void setCategoryRepository(CategoryRepository categoryRepository) {
         this.categoryRepository = categoryRepository;
+    }
+
+    @Autowired
+    public void setProductRepository(ProductRepository productRepository) {
+        this.productRepository = productRepository;
     }
 
     @GetMapping("/index")
@@ -33,11 +45,21 @@ public class CategoryController {
         return mav;
     }
 
-    @PostMapping(value = {"/create", "/edit"})
-    public ModelAndView create(@ModelAttribute("category") Category category,
+    @PostMapping(value = "/create")
+    public ModelAndView create(@Valid @ModelAttribute("category") Category category,
+                               BindingResult result) {
+        if (result.hasErrors()) {
+            return new ModelAndView("/category/create");
+        }
+        categoryRepository.save(category);
+        return new ModelAndView("redirect:index");
+    }
+
+    @PostMapping(value = "/edit")
+    public ModelAndView edit(@Valid @ModelAttribute("category") Category category,
                                BindingResult result){
         if(result.hasErrors()){
-            return new ModelAndView("error");
+            return new ModelAndView("/category/edit");
         }
         categoryRepository.save(category);
         return new ModelAndView("redirect:index");
@@ -63,7 +85,10 @@ public class CategoryController {
     @GetMapping("/delete")
     public ModelAndView delete(@RequestParam("id") int id){
         ModelAndView mav = new ModelAndView("redirect:index");
-        categoryRepository.delete(id);
+        List<Product> productList = productRepository.findByCategoryId(id);
+        if (productList == null || productList.isEmpty()) {
+            categoryRepository.delete(id);
+        }
         return mav;
     }
